backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
google-auth
requests
psycopg2-binary
dj-database-url
gunicorn
django-cors-headers
qrcode
Pillow
python-dotenv
whitenoise
